refactor the azure storage blob client classes, in preparation for sharing the client code between public and private blob storage.
