+++ server/app/services/cloud/azure/BaseAzureBlobStorageClient.java
@@ -1,0 +1,26 @@
+package services.cloud.azure;
+
+import com.azure.storage.blob.BlobServiceClient;
+import com.azure.storage.blob.models.BlobCorsRule;
+import com.azure.storage.blob.models.BlobServiceProperties;
+import java.util.List;
+
+/** Implements common logic for the Azure blob storage client. */
+abstract class BaseAzureBlobStorageClient implements AzureBlobStorageClientInterface {
+  protected void setCorsRules(BlobServiceClient blobServiceClient, String baseUrl) {
+    BlobServiceProperties properties =
+        new BlobServiceProperties()
+            .setCors(
+                List.of(
+                    new BlobCorsRule()
+                        .setAllowedOrigins(baseUrl)
+                        // These headers are generated by the azure sdk
+                        // and passed along, more info on the headers:
+                        // https://docs.microsoft.com/en-us/rest/api/storageservices/put-blob
+                        .setAllowedHeaders(
+                            "content-type,x-ms-blob-type,x-ms-client-request-id,x-ms-version,x-ms-blob-content-type,x-ms-blob-content-disposition")
+                        .setAllowedMethods("GET,PUT,OPTIONS")
+                        .setMaxAgeInSeconds(500)));
+    blobServiceClient.setProperties(properties);
+  }
+}
